requirements.txt
discord
openai
python-dotenv
flask
astrapy
requests
Pillow
vaderSentiment>=3.3.2
nltk>=3.8.1